requirements.txt
httpx
pydantic
python-dotenv
beautifulsoup4
requests